+++ defaults.js
@@ -1,0 +1,4 @@
+export default {
+    reactive: false,
+    single: false,
+}