+++ grapher-react.js
@@ -1,3 +1,0 @@
-export {
-    default as createQueryContainer
-} from './lib/createQueryContainer.js';