+++ main.client.js
@@ -1,0 +1,11 @@
+export {
+    default as setDefaults
+} from './setDefaults.js';
+
+export {
+    default as withQuery
+} from './withQuery.js';
+
+export {
+    default as createQueryContainer
+} from './legacy/createQueryContainer.js';