+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'cultofcoders:grapher-react',
-  version: '0.0.5',
+  version: '0.1.0',
   // Brief, one-line summary of the package.
   summary: 'Provides easy to use React Components that are suitable for grapher package.',
   // URL to the Git repository containing the source code for this package.
@@ -13,11 +13,11 @@ Package.describe({
 Package.onUse(function(api) {
   api.versionsFrom('1.3');
   api.use('ecmascript');
-  api.use('react-meteor-data@0.2.9');
-  api.imply('react-meteor-data@0.2.9');
-  api.use('aldeed:simple-schema@1.5.3');
+  api.imply('react-meteor-data');
+  api.imply('cultofcoders:grapher');
 
-  api.mainModule('grapher-react.js', 'client');
+  api.mainModule('main.client.js', 'client');
+  api.mainModule('main.server.js', 'server');
 });
 
 Package.onTest(function(api) {