+++ setDefaults.js
@@ -1,0 +1,5 @@
+import defaults from './defaults';
+
+export default function setDefaults(newDefaults) {
+    Object.assign(defaults, newDefaults);
+}